kram - added KTXImageData to isolate the memory backing a KTXImage
There were too many steps to wrap mmap, file helper, and alias data from a bundle.
Switch FileHelper to size_t.  Used to be int, but that's limited to 32-bit.
Simplify kram setup code for dealing with files.  PNG is still not tied into KTXImage so has to replicate some things.
Bubbling KTXImage up and out of loading.  It's useful to decode blocks and reference in the viewer.

+++ libkram/kram/KramFileHelper.h
@@ -32,18 +32,18 @@ public:
 
     void close();
 
-    // returns -1 if stat call fails
-    int64_t size() const;
+    // returns (size_t)-1 if stat call fails
+    size_t size() const;
 
     FILE* pointer() { return _fp; }
 
     // safe calls that test bytes read/written
-    bool read(uint8_t* data, int dataSize);
-    bool write(const uint8_t* data, int dataSize);
+    bool read(uint8_t* data, size_t dataSize);
+    bool write(const uint8_t* data, size_t dataSize);
 
     // if caller only has FILE* then can use these
-    static bool readBytes(FILE* fp, uint8_t* data, int dataSize);
-    static bool writeBytes(FILE* fp, const uint8_t* data, int dataSize);
+    static bool readBytes(FILE* fp, uint8_t* data, size_t dataSize);
+    static bool writeBytes(FILE* fp, const uint8_t* data, size_t dataSize);
 
     // return mod stamp on filename
     static uint64_t modificationTimestamp(const char* filename);
